Add exp backoff for downloads, refactor utils
