+++ python/mxnet/gluon/data/vision/datasets.py
@@ -29,7 +29,7 @@ import warnings
 import numpy as np
 
 from .. import dataset
-from ...utils import download, check_sha1
+from .... utils import download, check_sha1
 from .... import nd, image, recordio
