+++ python/mxnet/gluon/model_zoo/model_store.py
@@ -22,7 +22,7 @@ __all__ = ['get_model_file', 'purge']
 import os
 import zipfile
 
-from ..utils import download, check_sha1
+from ... utils import download, check_sha1
 
 _model_sha1 = {name: checksum for checksum, name in [
     ('44335d1f0046b328243b32a26a4fbd62d9057b45', 'alexnet'),
