+++ tests/python/unittest/test_gluon_data.py
@@ -42,7 +42,7 @@ def prepare_record():
     if not os.path.isdir("data/test_images"):
         os.makedirs('data/test_images')
     if not os.path.isdir("data/test_images/test_images"):
-        gluon.utils.download("http://data.mxnet.io/data/test_images.tar.gz", "data/test_images.tar.gz")
+        mx.utils.download("http://data.mxnet.io/data/test_images.tar.gz", "data/test_images.tar.gz")
         tarfile.open('data/test_images.tar.gz').extractall('data/test_images/')
     if not os.path.exists('data/test.rec'):
         imgs = os.listdir('data/test_images/test_images')
